Refactor real_env: rename helpers to data_handler and simplify ModelClient lifecycle.
Eagerly connect ModelClient in RealEnv.__init__ and drop lazy property plus optional None checks.
